Fix problem with priority option when submit job #16 (#9)
* adding feature

* fix

+++ src/main/java/com/epam/grid/engine/entity/job/JobOptions.java
@@ -59,7 +59,7 @@ public class JobOptions {
     /**
      * Job priority.
      */
-    private int priority;
+    private Long priority;
     /**
      * A list of queues in which the job should be processed.
      */
